% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARLfunctions.R
\name{FSL_read}
\alias{FSL_read}
\title{Read an FSL formatted radiosonde file}
\usage{
FSL_read(file_name, xlim = c(-180, 180), ylim = c(-90, 90))
}
\arguments{
\item{file_name}{name of the file to read}

\item{xlim}{left and right longitude limits to extract}

\item{ylim}{lower and upper latitude limits to extract}
}
\value{
a data table is returned with launch timestamp and location, altitude, pressure, temperature, and U and V wind components
}
\description{
Read an FSL formatted radiosonde file
}
\examples{
file_name <- 'roab_soundings43993.txt'
DT <- FSL_read(file_name,xlim=c(-125,-65),ylim=c(20,50))
sub_DT <- subset(DT, tstamp == DT$tstamp[1] )
plot(sub$lon, sub$lat)
map("worldHires",add=True)
}
